requirements.txt
# --- CORE & NETWORK ---
scapy>=2.5.0
requests>=2.28.0
aiohttp>=3.8.0
aiodns>=3.0.0
paramiko>=3.0.0

# --- DATABASE & ENV ---
aiosqlite>=0.18.0
python-dotenv>=0.21.0

# --- AI & MACHINE LEARNING (Критично для работы --update-ai) ---
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0

# --- UI & REPORTING ---
rich>=12.6.0
fpdf2>=2.7.0
chardet>=5.0.0

# --- DEPLOYMENT & TESTING ---
docker>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0

